refactor: Moved dark mode to its own component
