+++ pages/merchants.tsx
@@ -32,25 +32,6 @@ const Merchants: NextPage = (props) => {
     'merchantServer',
     'Shandi'
   )
-  const isMounted = useRef(false);
-  const defaultTheme = () => {
-    // Defaults to system theme if unconfigured
-    return (localStorage.getItem('darkMode') || window.matchMedia && window.matchMedia('(prefers-color-scheme: dark)').matches)
-  }
-  const [darkMode, setDarkMode] = useLocalStorage<boolean>('darkMode', defaultTheme)
-  useEffect(()=> {
-    //Prevents FoUC (Flash of Unstylized Content) by not refreshing on first mount
-    if (!isMounted.current){ isMounted.current = true; return }
-
-    //Toggle Daisy UI colors (e.g. bg-base-###)
-    document.documentElement.setAttribute('data-theme', darkMode ? 'dark' : 'light') 
-    
-    //Toggle standard Tailwind colors (e.g. bg-sky-800)
-    darkMode 
-      ?  document.documentElement.classList.add("dark")
-      :  document.documentElement.classList.remove("dark")
-  }, [darkMode])
-
   const [currDate, setCurrDate] = useState<DateTime>(DateTime.now())
   const [regionTZ, setRegionTZ] = useLocalStorage<string>('regionTZ', 'UTC-7')
